requirements.txt
anthropic>=0.40.0
duckdb>=1.1.0
pandas>=2.2.0
pyarrow>=17.0.0
pyyaml>=6.0.2
python-dotenv>=1.0.0
